+++ day11/src/operation.rs
@@ -2,14 +2,14 @@
 pub enum Operation {
     #[default]
     Noop,
-    Add(u32),
-    Multiply(u32),
+    Add(u64),
+    Multiply(u64),
     AddSelf,
     MultiplySelf,
 }
 
 impl Operation {
-    pub fn calc(&self, val: u32) -> u32 {
+    pub fn calc(&self, val: u64) -> u64 {
         match self {
             Operation::Noop => val,
             Operation::Add(n) => val + n,
